# server/src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WoxMail Master Suite REST & Streaming API",
    "version": "1.0.0",
    "description": "Complete, production-grade REST API reference for WoxMail: permanent mailboxes, temporary throwaway disposable mail, Enclave Vault confidential messages, Ephemeral Vector Streams, Time Capsule Future Letters, The Gatekeeper Cold Email Screener, WoxSMS, Developer CLI Gateway, Application Passwords, and Admin Management.",
    "contact": {
      "name": "WoxMail Engineering",
      "url": "https://mail.wox.world",
      "email": "support@wox.world"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://mail.wox.world/terms"
    }
  },
  "servers": [
    {
      "url": "https://mail.wox.world",
      "description": "Production Live Server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Standard JWT Bearer token obtained from `/api/auth/login`."
      },
      "appPasswordAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Developer SMTP / Application Password token formatted as `wox_app_xxxx-xxxx-xxxx-xxxx`."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Developer CLI / API Key header for headless scripts and daemons."
      },
      "tempSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "woxmail_temp",
        "description": "Encrypted session cookie for temporary disposable mailboxes."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token",
        "description": "CSRF token required for state-mutating requests (POST, PUT, DELETE, PATCH)."
      }
    },
...
  },
  "paths": {
...
    "/api/campaigns": {
      "get": {
        "summary": "List Broadcast Campaigns",
        "description": "Retrieves all newsletter and marketing email campaigns with delivery telemetry.",
        "tags": [
          "Broadcast Campaigns"
        ],
        "responses": {
          "200": {
            "description": "Campaigns list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "campaigns": []
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create and Schedule Campaign",
        "description": "Creates a mass broadcast campaign with template variable interpolation and unsubscribe handling.",
        "tags": [
          "Broadcast Campaigns"
        ],
        "responses": {
          "201": {
            "description": "Campaign created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": 1,
                    "status": "scheduled"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "required": true
                  },
                  "subject": {
                    "type": "string",
                    "required": true
                  },
                  "htmlContent": {
                    "type": "string",
                    "required": true
                  },
                  "recipientListId": {
                    "type": "integer",
                    "required": true
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}